requirements.txt
# =============================================================================
# ML Pipeline Requirements (Production)
# =============================================================================

# Core Data Science Libraries
pandas>=1.5.0
numpy>=1.23.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.2.0
xgboost>=1.7.0

# Model Persistence
joblib>=1.2.0

# MLflow for Model Registry & Tracking
mlflow>=2.8.0

# API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Dashboard
streamlit>=1.28.0
requests>=2.31.0

# Configuration & Environment
pyyaml>=6.0
python-dotenv>=1.0.0

# Database Connectors (Optional - install as needed)
# psycopg2-binary>=2.9.0  # PostgreSQL
# pymysql>=1.0.0          # MySQL

# Jupyter Notebook (for development)
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.19.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
